Added ci tests

+++ modules/typescript-express-module-example/jest.global-teardown.js
@@ -1,7 +1,10 @@
 const shell = require("shelljs");
 
 module.exports = async () => {
-  shell.exec("docker-compose -f docker-compose.test.yml down");
-  shell.exec("rm -r postgres-data-test");
-  shell.exec("docker system prune -f");
+  if (!process.env.CI) {
+    shell.exec("docker-compose -f docker-compose.test.yml down");
+    shell.exec("rm -r postgres-data-test");
+    shell.exec("docker system prune -f");
+  }
+  return;
 };
